Plugins now get auto included
