+++ src/ConsoleApp.SourceGenerator/ConsoleApp.SourceGenerator.csproj
@@ -11,8 +11,7 @@
 		<PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.0.1" PrivateAssets="all" />
 		<PackageReference Include="Microsoft.CodeAnalysis.Analyzers" Version="3.3.3" PrivateAssets="all" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-		<ProjectReference Include="..\SourceGenerator.Foundations.Contracts\SourceGenerator.Foundations.Contracts.csproj" ReferenceOutputAssembly="true"/>
-		<ProjectReference Include="..\SourceGenerator.Foundations\SourceGenerator.Foundations.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
+		<ProjectReference Include="..\SourceGenerator.Foundations.Contracts\SourceGenerator.Foundations.Contracts.csproj" ReferenceOutputAssembly="true" />
 	</ItemGroup>
 	<Import Project="..\SourceGenerator.Foundations\SourceGenerator.Foundations.props" />
 </Project>
