+++ src/Plugins/SourceGenerator.Foundations.Windows/Diagnostics/ProcessExtensions.cs
