+++ src/Plugins/SourceGenerator.Foundations.Windows/Diagnostics/ProcessInformation.cs
