--- src/SourceGenerator.Foundations.Windows/SourceGenerator.Foundations.VisualStudio.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<Nullable>enable</Nullable>
-		<LangVersion>10</LangVersion>
-		<RootNamespace>SGF</RootNamespace>
-		<TargetFramework>netstandard2.0</TargetFramework>
-	</PropertyGroup>
-	<ItemGroup>
-	  <ProjectReference Include="..\SourceGenerator.Foundations.Contracts\SourceGenerator.Foundations.Contracts.csproj" />
-		<PackageReference Include="envdte" Version="17.3.32804.24" />
-		<PackageReference Include="Microsoft.VisualStudio.Interop" Version="17.3.32804.24" GeneratePathProperty="true" PrivateAssets="all" />
-		<PackageReference Include="System.Management" Version="6.0.0" GeneratePathProperty="true" />
-	</ItemGroup>
-</Project>
