+++ src/SourceGenerator.Foundations.sln
@@ -11,12 +11,14 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleApp.SourceGenerator"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SourceGenerator.Foundations.Contracts", "SourceGenerator.Foundations.Contracts\SourceGenerator.Foundations.Contracts.csproj", "{E6A4D948-7741-4A56-92A4-C8B800CC7DDE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SourceGenerator.Foundations.VisualStudio", "SourceGenerator.Foundations.Windows\SourceGenerator.Foundations.VisualStudio.csproj", "{380588C0-0D48-49C9-AC2F-0A319F3FA28B}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{9DE06F12-24AF-41EF-8214-F015EEF952BB}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Plugins", "Plugins", "{A651676B-FBDB-4710-B010-D05AF3A56084}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SourceGenerator.Foundations.Windows", "Plugins\SourceGenerator.Foundations.Windows\SourceGenerator.Foundations.Windows.csproj", "{DC8C5A1A-6269-4BA7-852A-D21CB0B2B5A0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -40,13 +42,16 @@ Global
 		{E6A4D948-7741-4A56-92A4-C8B800CC7DDE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E6A4D948-7741-4A56-92A4-C8B800CC7DDE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E6A4D948-7741-4A56-92A4-C8B800CC7DDE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{380588C0-0D48-49C9-AC2F-0A319F3FA28B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{380588C0-0D48-49C9-AC2F-0A319F3FA28B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{380588C0-0D48-49C9-AC2F-0A319F3FA28B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{380588C0-0D48-49C9-AC2F-0A319F3FA28B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DC8C5A1A-6269-4BA7-852A-D21CB0B2B5A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DC8C5A1A-6269-4BA7-852A-D21CB0B2B5A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DC8C5A1A-6269-4BA7-852A-D21CB0B2B5A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DC8C5A1A-6269-4BA7-852A-D21CB0B2B5A0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{DC8C5A1A-6269-4BA7-852A-D21CB0B2B5A0} = {A651676B-FBDB-4710-B010-D05AF3A56084}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {EDB10920-970A-43F9-A2B3-7F1270DD477B}
